requirements.txt
argcomplete==1.10.0
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
defusedxml==0.6.0
Django==2.2.12
django-currentuser==0.5.0
django-summernote==0.8.11.6
django-tinymce==2.8.0
djangorestframework==3.11.0
EbookLib==0.17.1
et-xmlfile==1.0.1
extract-msg==0.23.1
gunicorn==20.0.4
IMAPClient==2.1.0
isort==4.3.21
jdcal==1.4.1
joblib==0.14.1
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
odfpy==1.4.0
olefile==0.46
pdfminer.six==20181108
Pillow==7.0.0
pipenv==2018.11.26
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycryptodome==3.9.7
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.14.0
sortedcontainers==2.1.0
sqlparse==0.3.1
tablib==0.14.0
typed-ast==1.4.0
tzlocal==1.5.1
urllib3==1.25.7
virtualenv==16.6.2
virtualenv-clone==0.5.3
whitenoise==5.0.1
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
